made RoseTree to a new module and changed name from shrink to RoseGen
